feat(examples): pull discovery and gossip into examples
