+++ examples/README.md
@@ -1,0 +1,17 @@
+## Examples
+
+These examples demonstrate how to work with kona crates.
+Some examples are isolated services broken out from OP Stack components.
+
+To run an example, use the command `cargo run -p <example>`.
+
+If you have an idea for a new example, [open an issue][issue].
+Otherwise if you already have an example you'd like to add, open a PR!
+
+#### Discovery
+
+
+
+<!-- Links -->
+
+[issue]: https://github.com/op-rs/kona/issues/new
