+++ examples/discovery/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "example-discovery"
+version = "0.0.0"
+publish = false
+edition.workspace = true
+license.workspace = true
+
+[dependencies]
+anyhow.workspace = true
+tracing.workspace = true
+kona-cli.workspace = true
+kona-p2p.workspace = true
+clap = { workspace = true, features = ["derive", "env"] }
+tokio = { workspace = true, features = ["rt-multi-thread", "macros"] }
+tracing-subscriber = { workspace = true, features = ["fmt", "env-filter"] }
